+++ test/integration/BasketTest.php
@@ -82,7 +82,8 @@ class BasketTest extends BasePaymentTest
             ->setVat(19)
             ->setUnit('ert')
             ->setAmountDiscount(1234.9)
-            ->setImageUrl('https://files.readme.io/9f556bd-small-Heidelpay-Logo_mitUnterzeile-orange_RGB.jpg');
+            ->setImageUrl('https://files.readme.io/9f556bd-small-Heidelpay-Logo_mitUnterzeile-orange_RGB.jpg')
+            ->setSubTitle('This is some subtitle for this item');
         $basket->addBasketItem($basketItem);
         $this->assertEmpty($basket->getId());