+++ CHANGELOG.md
@@ -15,7 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 *   Updated unit tests.
 *   Refactored examples.
 *   Allow to pass the event payload manually to `Heidelpay::fetchResourceFromEvent`.
-*   Add property `imageUrl` to `BasketItem` resource.
+*   Add properties `imageUrl` and `subTitle` to `BasketItem` resource.
 
 ## [1.1.3.0][1.1.3.0]